Move widget setup code to widget module
