+++ core/build.gradle
@@ -39,6 +39,7 @@ dependencies {
     implementation project(':storage:preferences')
     implementation project(':ui:app-start-intent')
     implementation project(':ui:common')
+    implementation project(':ui:episodes')
     implementation project(':ui:i18n')
     implementation project(':ui:notifications')
     implementation project(':ui:widget')
