+++ settings.gradle
@@ -40,6 +40,7 @@ include ':storage:preferences'
 include ':ui:app-start-intent'
 include ':ui:common'
 include ':ui:echo'
+include ':ui:episodes'
 include ':ui:glide'
 include ':ui:i18n'
 include ':ui:notifications'
