+++ ui/common/src/main/res/values/dimens.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="widget_margin">0dp</dimen>
-    <dimen name="widget_inner_radius">4dp</dimen>
     <dimen name="external_player_height">64dp</dimen>
     <dimen name="text_size_micro">12sp</dimen>
     <dimen name="text_size_small">14sp</dimen>
