+++ ui/echo/build.gradle
@@ -18,6 +18,7 @@ dependencies {
     implementation project(':storage:database')
     implementation project(":storage:preferences")
     implementation project(':ui:common')
+    implementation project(':ui:episodes')
     implementation project(':ui:glide')
 
     annotationProcessor "androidx.annotation:annotation:$annotationVersion"
