+++ ui/episodes/README.md
@@ -1,0 +1,3 @@
+# :ui:episodes
+
+Common classes that are needed everywhere we display information about episodes.
