+++ ui/episodes/src/main/java/de/danoeh/antennapod/ui/episodes/ImageResourceUtils.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.core.feed.util;
+package de.danoeh.antennapod.ui.episodes;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
