+++ ui/widget/build.gradle
@@ -2,6 +2,7 @@ plugins {
     id("com.android.library")
 }
 apply from: "../../common.gradle"
+apply from: "../../playFlavor.gradle"
 
 android {
     namespace "de.danoeh.antennapod.ui.widget"
@@ -10,10 +11,25 @@ android {
         vectorDrawables.useSupportLibrary false
         vectorDrawables.generatedDensities = ["xhdpi"]
     }
+
+    lint {
+        disable "IconMissingDensityFolder"
+    }
 }
 
 dependencies {
+    implementation project(":model")
+    implementation project(":playback:base")
+    implementation project(':storage:preferences')
+    implementation project(':storage:database')
+    implementation project(":ui:app-start-intent")
     implementation project(":ui:common")
+    implementation project(":ui:episodes")
+    implementation project(':ui:glide')
+    implementation project(':ui:i18n')
 
     annotationProcessor "androidx.annotation:annotation:$annotationVersion"
+    implementation "androidx.appcompat:appcompat:$appcompatVersion"
+    implementation "androidx.work:work-runtime:$workManagerVersion"
+    implementation "com.github.bumptech.glide:glide:$glideVersion"
 }
