+++ ui/widget/src/main/res/drawable-hdpi/ic_widget_preview.png
